Rename to "manual_codegen_test"
